Add Random information section
Add random information section on the website that includes a header and
four pictures with some text. Add the pictures to the folder and the
appropriate styling. Apply some minor changes to the header's section
css and more specifically switched to paddings from margins to retain
the background color at the section transition.